+++ gh-pages/changelog.md
@@ -1,3 +1,10 @@
+---
+title: Change Log
+permalink: /changelog
+---
+
+{% assign package = site.data.package %}
+
 # Change Log
 
 All notable changes to this project will be documented in this file.
@@ -6,6 +13,9 @@ All notable changes to this project will be documented in this file.
 Changelog format adheres to [Keep a Changelog](http://keepachangelog.com/)</sub>
 
 ## [Unreleased]
+### Added
+- Everything is a Tensor
+- Composition algebra class
 
 ## [0.7.0] - 2016-03-09
 ### Added
@@ -15,5 +25,5 @@ Changelog format adheres to [Keep a Changelog](http://keepachangelog.com/)</sub>
 - first tensro implementation
 - minified code and source map
 
-[Unreleased]: https://github.com/fibo/algebra/compare/v0.6.2...HEAD
-[0.7.0]: https://github.com/fibo/algebra/compare/v0.6.2...v0.7.0
+[Unreleased]: https://github.com/fibo/{{ package.name }}/compare/v0.6.2...HEAD
+[0.7.0]: https://github.com/fibo/algebra/{{ package.name }}/v0.6.2...v0.7.0